@@ -2,6 +2,8 @@
 
 namespace App\Entity;
 
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 
 /**
@@ -19,17 +21,45 @@ class Product
     /**
      * @ORM\Column(type="string", length=255)
      */
-    private $name;
+    private string $name;
 
     /**
      * @ORM\Column(type="string", length=255)
      */
-    private $image;
+    private string $image;
 
     /**
      * @ORM\Column(type="decimal", precision=10, scale=2)
      */
-    private $price;
+    private string $price;
+
+    /**
+     * @ORM\OneToMany(
+     *     targetEntity="App\Entity\Feedback",
+     *     mappedBy="product",
+     *     orphanRemoval=true,
+     *     cascade={"persist"}
+     * )
+     */
+    private Collection $feedbacks;
+
+    /**
+     * @ORM\ManyToMany(targetEntity="App\Entity\CharacteristicValue")
+     * @ORM\JoinTable(
+     *     name="products_characteristic_values",
+     *     joinColumns={@ORM\JoinColumn(name="product_id", nullable=false)},
+     *     inverseJoinColumns={
+     *         @ORM\JoinColumn(name="characteristic_value_id", nullable=false)
+     *     }
+     * )
+     */
+    private Collection $characteristicValues;
+
+    public function __construct()
+    {
+        $this->feedbacks            = new ArrayCollection();
+        $this->characteristicValues = new ArrayCollection();
+    }
 
     public function getId(): ?int
     {
@@ -68,6 +98,51 @@ class Product
     public function setPrice(string $price): self
     {
         $this->price = $price;
+
+        return $this;
+    }
+
+    /**
+     * @return Collection|Feedback[]
+     */
+    public function getFeedbacks(): Collection
+    {
+        return $this->feedbacks;
+    }
+
+    public function addFeedback(Feedback $feedback): self
+    {
+        if (!$this->feedbacks->contains($feedback)) {
+            $this->feedbacks[] = $feedback;
+            $feedback->setProduct($this);
+        }
+
+        return $this;
+    }
+
+    public function removeFeedback(Feedback $feedback): self
+    {
+        if ($this->feedbacks->contains($feedback)) {
+            $this->feedbacks->removeElement($feedback);
+            // set the owning side to null (unless already changed)
+            if ($feedback->getProduct() === $this) {
+                $feedback->setProduct(null);
+            }
+        }
+
+        return $this;
+    }
+
+    public function getCharacteristicValues(): Collection
+    {
+        return $this->characteristicValues;
+    }
+
+    public function addCharacteristicValue(CharacteristicValue $characteristicValue): self
+    {
+        if (!$this->characteristicValues->contains($characteristicValue)) {
+            $this->characteristicValues->add($characteristicValue);
+        }
 
         return $this;
     }
